refactor: 익명 여부에 따른 프로필 이미지 컴포넌트 안에서 처리하도록 변경
